+++ src/index.js
@@ -4,7 +4,7 @@ import { createStore, applyMiddleware } from 'redux';
 import thunk from 'redux-thunk';
 import { Provider } from 'react-redux';
 import 'semantic-ui-css/semantic.min.css';
-import App from './components/App';
+import AppContainer from './containers/AppContainer';
 import sandglass from './reducers'
 import registerServiceWorker from './config/registerServiceWorker';
 
@@ -12,7 +12,7 @@ let store = createStore(sandglass, applyMiddleware(thunk));
 
 render(
   <Provider store={store}>
-    <App />
+    <AppContainer />
   </Provider>,
   document.getElementById('root')
 );
